+++ src/components/drawers/RResearch.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>Research Drawer</div>
+</template>
+
+<script>
+export default {
+  // name: 'ComponentName',
+  data () {
+    return {}
+  }
+}
+</script>
+
+<style>
+</style>